# web/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NextQuestion": {
        "properties": {
          "question": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QuestionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "waiting": {
            "default": false,
            "description": "Questions about the memory just recorded are still on their way: ask again in a moment",
            "title": "Waiting",
            "type": "boolean"
          }
        },
        "required": [
          "question"
        ],
        "title": "NextQuestion",
        "type": "object"
      },
...
      "Page_QuestionOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QuestionOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[QuestionOut]",
        "type": "object"
      },
...
      "QuestionIn": {
        "properties": {
          "event_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "period_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Id"
          },
          "person_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Id"
          },
          "scope": {
            "$ref": "#/components/schemas/Scope",
            "default": "general"
          },
          "text": {
            "maxLength": 1000,
            "minLength": 3,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "QuestionIn",
        "type": "object"
      },
      "QuestionOut": {
        "properties": {
          "about": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "What it asks about: the event, period, person or story",
            "title": "About"
          },
          "answered_by_memory_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answered By Memory Id"
          },
          "asked_of_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asked Of Id"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "period_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Id"
          },
          "person_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Id"
          },
          "scope": {
            "$ref": "#/components/schemas/Scope"
          },
          "source_memory_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Memory Id"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "text",
          "status",
          "scope",
          "source_memory_id",
          "event_id",
          "period_id",
          "person_id",
          "answered_by_memory_id"
        ],
        "title": "QuestionOut",
        "type": "object"
      },
      "QuestionPatch": {
        "properties": {
          "answered_by_memory_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answered By Memory Id"
          },
          "status": {
            "description": "A pending question is answered or dismissed, for good",
            "enum": [
              "answered",
              "dismissed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "QuestionPatch",
        "type": "object"
      },
...
      "Scope": {
        "enum": [
          "general",
          "period",
          "event",
          "person"
        ],
        "title": "Scope",
        "type": "string"
      },
...
      "Status": {
        "enum": [
          "pending",
          "answered",
          "dismissed"
        ],
        "title": "Status",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Memoir API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/questions": {
      "get": {
        "operationId": "list_questions_api_questions_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Status",
              "default": "pending"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Scope"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scope"
            }
          },
          {
            "in": "query",
            "name": "event_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Id"
            }
          },
          {
            "in": "query",
            "name": "person_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Person Id"
            }
          },
          {
            "description": "Raised by this memory",
            "in": "query",
            "name": "source_memory_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Raised by this memory",
              "title": "Source Memory Id"
            }
          },
          {
            "description": "Answered by it",
            "in": "query",
            "name": "answered_by_memory_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Answered by it",
              "title": "Answered By Memory Id"
            }
          },
          {
            "description": "From the previous page",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "From the previous page",
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_QuestionOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Questions",
        "tags": [
          "questions"
        ]
      },
      "post": {
        "operationId": "ask_api_questions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask",
        "tags": [
          "questions"
        ]
      }
    },
...
    "/api/questions/next": {
      "get": {
        "description": "The one question to ask now, for the home screen and after each recording.",
        "operationId": "next_question_api_questions_next_get",
        "parameters": [
          {
            "description": "The memory just recorded, whose questions should come first",
            "in": "query",
            "name": "after_memory_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The memory just recorded, whose questions should come first",
              "title": "After Memory Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextQuestion"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Next Question",
        "tags": [
          "questions"
        ]
      }
    },
    "/api/questions/{question_id}": {
      "delete": {
        "operationId": "remove_question_api_questions__question_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Question",
        "tags": [
          "questions"
        ]
      },
      "get": {
        "operationId": "one_question_api_questions__question_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One Question",
        "tags": [
          "questions"
        ]
      },
      "patch": {
        "operationId": "answer_or_dismiss_api_questions__question_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Answer Or Dismiss",
        "tags": [
          "questions"
        ]
      }
    },
...
  }
}